Fixed struct bug from missing types
IDA contains types in its typelibs that it doesn't export using
IDAPython API's. If an exception is thrown when defining a struct
members type, I re-type it to be a byte array. Also, I've added
OrderedDict's to structure dictionaries in hope that it prevents types
from being applied out of order.